+++ Math/Math/CPUSparseMatrix.h
@@ -79,7 +79,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
         
         int GetComputeDeviceId() const {return -1;}
         
-        void Resize(const size_t numRows, const size_t numCols, int size = 0);
+        void Resize(const size_t numRows, const size_t numCols, size_t size = 0);
         void Reset();
 
     public:
